gourt literacy programme (#9)
* simplify Word and add icon

* implement basic framework for saying words in need of fleshing out. Fall back on exisitng interact framework that than make bespoke examing framework since they are basically the same.

* Add the code fo spawning the speech bubble and lay the groundwork for speech bubbles being easier to work with in the future. And fix a few bugs and problems I ran into in the process.
Change get_bounds to get_rect for better consistency, and deal with some complications stemming from the fact that they don't work exactly the same. Fix two bugs in speech_bubble, one being incorrect initialization and the other being a faulty assumption that a Rect2 will always be centred on the owner.
Remove `examinable` layer since examination is done via interaction and is therefore on the `interactive` layer

* remove rogue print

* fix another bug caused by assuming a rect would be centred over its owner. Add some new debug stuff

* Offload managing timing out the speech bubble to the speech bubble itself. Change and simplify some interfaces accordingly. Add settings.gd to hold global text speed.
Append text to existing speech bubbles instead of overwriting when `say` is invoked.

* add conversation and actors back into the office level

---------

Co-authored-by: jcomcl <j@jcom.cl>